+++ server/server.go
@@ -37,7 +37,6 @@ import (
 	"github.com/runatlantis/atlantis/server/events/locking/boltdb"
 	"github.com/runatlantis/atlantis/server/events/models"
 	"github.com/runatlantis/atlantis/server/events/run"
-	"github.com/runatlantis/atlantis/server/events/runtime"
 	"github.com/runatlantis/atlantis/server/events/terraform"
 	"github.com/runatlantis/atlantis/server/events/vcs"
 	"github.com/runatlantis/atlantis/server/events/webhooks"
@@ -201,10 +200,11 @@ func NewServer(userConfig UserConfig, config Config) (*Server, error) {
 		ConfigReader: configReader,
 		Terraform:    terraformClient,
 	}
-	executionPlanner := &runtime.ExecutionPlanner{
+	executionPlanner := &events.ExecutionPlanner{
 		ParserValidator:   &yaml.ParserValidator{},
 		DefaultTFVersion:  terraformClient.Version(),
 		TerraformExecutor: terraformClient,
+		ProjectFinder:     &events.DefaultProjectFinder{},
 	}
 	underlyingRouter := mux.NewRouter()
 	router := &Router{
@@ -289,6 +289,8 @@ func NewServer(userConfig UserConfig, config Config) (*Server, error) {
 		RepoWhitelist:          repoWhitelist,
 		SupportedVCSHosts:      supportedVCSHosts,
 		VCSClient:              vcsClient,
+		AtlantisGithubUser:     models.User{Username: userConfig.GithubUser},
+		AtlantisGitlabUser:     models.User{Username: userConfig.GitlabUser},
 	}
 	return &Server{
 		AtlantisVersion:    config.AtlantisVersion,